requirements.txt
spotipy==2.25.1
python-dotenv==1.2.1